development/code/phase4/palmer-living-reality/server/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.5.1
SQLAlchemy==2.0.23
SQLAlchemy-Serializer==1.4.12
python-dotenv==1.0.0
Werkzeug==3.0.1
requests==2.31.0
python-dateutil==2.8.2

